% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.r
\name{evaluateTask}
\alias{evaluateTask}
\title{Evaluate one iteration from a rocto job.}
\usage{
evaluateTask(roctoJob, outputDir, iterId = "test")
}
\arguments{
\item{roctoJob}{<character> Path to a \code{.rocto} file containing the job.}

\item{outputDir}{<character> Full directory where the output file should be saved.}

\item{iterId}{<integer> or <character: "test"> Indicates which task \code{[1:nIter]}
should be run, or whether the test parameters should be used.}
}
\value{
Invisible boolean TRUE.
}
\description{
This function runs one rocto job iteration - a "task".
}
\seealso{
\code{\link{newJob}}, \code{\link{packJob}}, \code{\link{resultsToList}}
}
