% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{newJob}
\alias{newJob}
\title{Initialise a new rocto job}
\usage{
newJob(name = "roctoJob", path = ".")
}
\arguments{
\item{name}{<character> Name to be given to the new job directory}

\item{path}{<character> Where to create the directory (default to current working directory)}
}
\value{
Invisible boolean TRUE
}
\description{
Creates a skeleton for a new \code{rocto} job in the correct format. Also allows you
to automatically set the working directory and open the param and main files.
}
\seealso{
\code{\link{packJob}}
}
