% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{roctoCheck}
\alias{roctoCheck}
\title{Check a rocto job}
\usage{
roctoCheck(path = ".")
}
\arguments{
\item{path}{<character> Path to existing \code{rocto} directory (default to current working directory)}
}
\value{
Invisible boolean valid job or not.
}
\description{
Checks whether an existing \code{rocto} job contains any errors.
It also specifies where problems may be and warns about unexpected deviations
from the default \code{rocto} folder format.
}
\seealso{
\code{\link{roctoNew}} \code{\link{roctoPack}}
}
