% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{roctoNew}
\alias{roctoNew}
\title{Initialise a new rocto job}
\usage{
roctoNew(name = "roctoJob", path = ".", interactive = TRUE)
}
\arguments{
\item{name}{<character> Name to be given to the new job directory}

\item{path}{<character> Where to create the directory (default to current working directory)}

\item{interactive}{<boolean> Whether to ask for feedback from user.}
}
\value{
Invisible boolean TRUE
}
\description{
Creates a skeleton for a new \code{rocto} job in the correct format. Also allows you
to automatically set the working directory and open the param and main files.
}
\seealso{
\code{\link{roctoPack}}
}
