% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{roctoPack}
\alias{roctoPack}
\title{Pack a rocto directory for distribution}
\usage{
roctoPack(path = ".", verbose = FALSE)
}
\arguments{
\item{path}{<character> The root of the raw rocto job directory.}

\item{verbose}{<boolean> Print debug information for inspecting parameter grid and meta info}
}
\value{
Invisible boolean TRUE
}
\description{
This function performs checks and then packs a raw \code{rocto} job directory 
into a single file for uploading to the \code{rocto} volunteer cluster.
}
\seealso{
\code{\link{roctoNew}}, \code{\link{roctoResults}}
}
