% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{roctoResults}
\alias{roctoResults}
\title{Load results from your rocto job into R}
\usage{
roctoResults(roctoResults)
}
\arguments{
\item{roctoResults}{<character> Path to a rocto results folder.}
}
\value{
List with \code{nIter} elements, each containing the results object of one iteration.
}
\description{
This function loads all the job results from a results folder into R at once.
}
\seealso{
\code{\link{roctoNew}}, \code{\link{roctoPack}}
}
